+++ resources/js/Icons/TikTokIcon.tsx
