+++ resources/js/Layouts/GuestLayout/index.tsx
