+++ resources/js/Layouts/GuestLayout/styles.tsx
