+++ resources/js/ziggy.d.ts
@@ -1,11 +1,15 @@
 /* This file is generated by Ziggy. */
-declare module "ziggy-js" {
-    interface RouteList {
-        "sanctum.csrf-cookie": [];
-        "ignition.healthCheck": [];
-        "ignition.executeSolution": [];
-        "ignition.updateConfig": [];
-        home: [];
-    }
+declare module 'ziggy-js' {
+  interface RouteList {
+    "sanctum.csrf-cookie": [],
+    "ignition.healthCheck": [],
+    "ignition.executeSolution": [],
+    "ignition.updateConfig": [],
+    "home": [],
+    "about": [],
+    "safety": [],
+    "support": [],
+    "careers": []
+}
 }
 export {};
