+++ resources/js/ziggy.js
@@ -1,29 +1,6 @@
-const Ziggy = {
-    url: "http://localhost",
-    port: null,
-    defaults: {},
-    routes: {
-        "sanctum.csrf-cookie": {
-            uri: "sanctum/csrf-cookie",
-            methods: ["GET", "HEAD"],
-        },
-        "ignition.healthCheck": {
-            uri: "_ignition/health-check",
-            methods: ["GET", "HEAD"],
-        },
-        "ignition.executeSolution": {
-            uri: "_ignition/execute-solution",
-            methods: ["POST"],
-        },
-        "ignition.updateConfig": {
-            uri: "_ignition/update-config",
-            methods: ["POST"],
-        },
-        home: { uri: "/", methods: ["GET", "HEAD"] },
-    },
-};
+const Ziggy = {"url":"http:\/\/localhost","port":null,"defaults":{},"routes":{"sanctum.csrf-cookie":{"uri":"sanctum\/csrf-cookie","methods":["GET","HEAD"]},"ignition.healthCheck":{"uri":"_ignition\/health-check","methods":["GET","HEAD"]},"ignition.executeSolution":{"uri":"_ignition\/execute-solution","methods":["POST"]},"ignition.updateConfig":{"uri":"_ignition\/update-config","methods":["POST"]},"home":{"uri":"\/","methods":["GET","HEAD"]},"about":{"uri":"about","methods":["GET","HEAD"]},"safety":{"uri":"safety","methods":["GET","HEAD"]},"support":{"uri":"support","methods":["GET","HEAD"]},"careers":{"uri":"careers","methods":["GET","HEAD"]}}};
 
-if (typeof window !== "undefined" && typeof window.Ziggy !== "undefined") {
+if (typeof window !== 'undefined' && typeof window.Ziggy !== 'undefined') {
     Object.assign(Ziggy.routes, window.Ziggy.routes);
 }
